+++ docs/main/changelog.rst
@@ -11,6 +11,9 @@ Development version
 -------------------
 - Add possibility to validate ``Factory`` provided type on ``Factory`` 
   initialization.
+- Add possibility to validate ``Singleton`` provided type on ``Singleton`` 
+  initialization.
+- Make some refactorings for providers.
 
 1.11.2
 ------